pyproject.toml
[project]
name = "SASTBenchmark"
version = "1.1.0"
description = "Scripts and wrappers for SASTs and datasets, with abstractions to support integration, benchmarking, and testing."
readme = "README.md"
license = "AGPL-3.0-only"
license-files = ["LICENSE"]
requires-python = ">=3.12"
authors = [
  {name = "Villon Chen"},
]
dependencies = [
    "blessed>=1.21.0",
    "click>=8.2.1",
    "gitpython>=3.1.44",
    "humanize>=4.12.3",
    "inquirer>=3.4.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
    "xmltodict>=0.14.2",
]

[project.scripts]
sastb = "sastbenchmark.main:cli"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sastbenchmark", "sastbenchmark.*"]

[tool.setuptools]
include-package-data = true

[tool.ruff.lint]
select = ["E", "F", "I", "ANN", "B", "TC"]
ignore = ["E501"]
